Moved video component test bed to different file
