+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/components/video-view-base/video-view-base.component.spec.ts
@@ -1,55 +1,14 @@
-import { Type, Component, Input } from '@angular/core';
 import { VideoUrlService } from '../../services/video-url.service';
-import { Config } from '../../../shared/models/config';
-import { Logger } from 'src/app/services/logger';
-import { MediaService } from '../../services/media.service';
-import { UserMediaService } from '../../services/user-media.service';
-import { UserCameraViewComponent } from './../../components/user-camera-view/user-camera-view.component';
 import { VideoViewBaseComponent } from './video-view-base.component';
-import { IndividualJourney } from '../../individual-journey';
 import { VideoFiles } from '../../services/video-files';
-import { IndividualStepsOrderFactory } from '../../individual-steps-order.factory';
-import { DeviceType } from '../../services/device-type';
-import { ComponentFixture } from '@angular/core/testing';
-import { IndividualJourneyComponentTestBed } from '../../pages/individual-base-component/individual-component-test-bed.spec';
-
-@Component({
-  selector: 'app-video-view',
-  template: ''
-})
-class StubVideoViewComponent {
-  @Input()
-  source: string;
-}
-const deviceType = jasmine.createSpyObj<DeviceType>(['isMobile']);
-
-export class VideoViewComponentTestBed {
-  static createComponent<TComponent>(component: Type<TComponent>): ComponentFixture<TComponent> {
-    return IndividualJourneyComponentTestBed.createComponent({
-      component: component,
-      providers: [
-        { provide: Logger, useValue: jasmine.createSpyObj<Logger>(['getVideoFileUrlerror']) },
-        { provide: VideoUrlService, useValue: jasmine.createSpyObj<VideoUrlService>(['getVideoFileUrl']) },
-        { provide: Config, useValue: {} },
-        { provide: MediaService, useClass: UserMediaService },
-        { provide: DeviceType, useValue: deviceType }
-      ],
-      declarations: [
-        StubVideoViewComponent,
-        UserCameraViewComponent
-      ]
-    });
-  }
-}
+import { IndividualJourneyStubs } from '../../pages/individual-base-component/individual-component-test-bed.spec';
 
 describe('functionality', () => {
   let component: VideoViewBaseComponent;
   const videoUrlService = jasmine.createSpyObj<VideoUrlService>(['getVideoFileUrl']);
-  const individualStepsOrderFactory = new IndividualStepsOrderFactory(deviceType);
-  deviceType.isMobile.and.returnValue(false);
 
   beforeEach(() => {
-    const journey = new IndividualJourney(individualStepsOrderFactory);
+    const journey = IndividualJourneyStubs.default;
     component = new VideoViewBaseComponent(journey, videoUrlService, VideoFiles.BeforeTheDay_Court);
   });
